Add optional systemd ready and watchdog support
If systemd is available, 2 new features are implemented:

1) systemd ready support: clustercheck sends now READY=1 via the
systemd notify socket when the socket is available. This is
useful to indicate to other services that clustercheck is ready to
accept connections.
To enable this feature, "Type=notify" must be set in the [Service]
section.
See also
https://www.freedesktop.org/software/systemd/man/sd_notify.html

2) systemd service watchdog support: clustercheck sends now WATCHDOG=1
via the systemd notify socket when the socket is available. This is
useful to make sure that clustercheck does not hang somewhere in the
process. In that case, the WATCHDOG ping would not be send and systemd
would restart the service automatically.
To enable this feature, "WatchdogSec=X" (where X are seconds) must be
set in the [Service] section.
See also
https://www.freedesktop.org/software/systemd/man/sd_watchdog_enabled.html